+++ test-apps/ember_3_10_mu/.eucrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  whitelist: ['z-button'],
+  ignore: [
+    'src/ui/routes/application/freestyle.hbs', // this is our template with style guides
+  ],
+};