+++ test-apps/ember_3_10_mu/src/ui/components/y-button/component.js
@@ -1,0 +1,9 @@
+import XButton from 'ember-3-10-mu/src/ui/components/x-button/component';
+
+export default XButton.extend({});
+
+/*
+  *** TEST CASE - find ***
+
+  Testing support for importing components and extending by other components.
+*/