+++ tests/ember_lts_3_4.js
@@ -8,6 +8,7 @@ test('3.4 LTS - get config', t => {
     appPath: '/test-apps/ember_lts_3_4/app/',
     ignore: ['app/templates/freestyle.hbs'],
     usePods: false,
+    useModuleUnification: false,
     whitelist: ['z-button'],
     componentsPath: '/test-apps/ember_lts_3_4/app/components',
   };
@@ -22,6 +23,7 @@ test('3.4 LTS - map components', t => {
     appPath: '/test-apps/ember_lts_3_4/app/',
     ignore: ['app/templates/freestyle.hbs'],
     usePods: false,
+    useModuleUnification: false,
     whitelist: ['z-button'],
     componentsPath: '/test-apps/ember_lts_3_4/app/components',
   };
@@ -73,6 +75,7 @@ test('3.4 LTS - look for unused components and calculate stats', t => {
     appPath: '/test-apps/ember_lts_3_4/app/',
     ignore: ['app/templates/freestyle.hbs'],
     usePods: false,
+    useModuleUnification: false,
     whitelist: ['z-button'],
     componentsPath: '/test-apps/ember_lts_3_4/app/components',
   };